Rearranged code and extracted cwm code into separate file
